+++ example/api_doc/flask_example.py
@@ -1,4 +1,5 @@
 from typing import Dict
+
 from example.param_verify.flask_example import create_app
 from pait.api_doc.markdown import PaitMd
 from pait.api_doc.open_api import PaitOpenApi
@@ -7,7 +8,6 @@ from pait.api_doc.pait_yaml import PaitYaml
 from pait.app import load_app
 from pait.model import PaitCoreModel
 
-
 if __name__ == "__main__":
     filename: str = "./example_doc/flask_pait"
     pait_dict: Dict[str, PaitCoreModel] = load_app(create_app())
@@ -15,4 +15,13 @@ if __name__ == "__main__":
     PaitJson(pait_dict, filename=filename, indent=2)
     PaitYaml(pait_dict, filename=filename)
     for i in ("json", "yaml"):
-        PaitOpenApi(pait_dict, title="Pait Doc", filename=filename + "_openapi", type_=i)
+        PaitOpenApi(
+            pait_dict,
+            title="Pait Doc",
+            open_api_tag_list=[
+                {"name": "test", "description": "test api"},
+                {"name": "user", "description": "user api"},
+            ],
+            type_=i,
+            filename=filename + "_openapi",
+        )