+++ pait/api_doc/markdown.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, List, Optional
 from pydantic.fields import Undefined
 
 from pait.model import PaitCoreModel, PaitStatus
+
 from .base_parse import PaitBaseParse  # type: ignore
 
 json_type_default_value_dict: Dict[str, Any] = {
@@ -24,7 +25,7 @@ class PaitMd(PaitBaseParse):
         pait_dict: Dict[str, PaitCoreModel],
         title: str = "Pait Doc",
         use_html_details: bool = True,
-        filename: Optional[str] = None
+        filename: Optional[str] = None,
     ):
         self._use_html_details: bool = use_html_details  # some not support markdown in html
         self._title: str = title
@@ -34,7 +35,21 @@ class PaitMd(PaitBaseParse):
         self.output(filename, markdown_text, ".md")
 
     @staticmethod
-    def gen_example_json(field_dict_list: List[dict], blank_num: int) -> str:
+    def gen_example_json(field_dict_list: List[dict], blank_num: int, indent: int = 2) -> str:
+        """
+        gen example json by field dict list
+        :param field_dict_list:
+        :param blank_num: Indent the first character of each line of json
+        :param indent: json indent
+        :return:
+        blank_num: -> .
+        indent: -> `
+        ....{
+        ....``"code": 0,
+        ....``"msg": "msg,
+        ....}
+        """
+        blank_num_str: str = " " * blank_num
         example_dict: dict = {}
         for field_dict in field_dict_list:
             sub_dict: dict = example_dict
@@ -49,13 +64,23 @@ class PaitMd(PaitBaseParse):
                         sub_dict[sub_key] = {}
                     sub_dict = sub_dict[sub_key]
             sub_dict[key] = json_type_default_value_dict[field_dict["type"]]
-        json_str: str = f"\n".join([" " * blank_num + i for i in json.dumps(example_dict, indent=2).split("\n")])
-        return f"{' ' * blank_num}```json\n" f"{json_str}\n" f"{' ' * blank_num}```\n\n"
+        json_str: str = f"\n".join([blank_num_str + i for i in json.dumps(example_dict, indent=indent).split("\n")])
+        return f"{blank_num_str}```json\n{json_str}\n{blank_num_str}```\n\n"
 
     @staticmethod
     def gen_md_param_table(field_dict_list: List[dict], blank_num: int = 8) -> str:
-        markdown_text: str = f"{' ' * blank_num}|param name|type|default value|description|other|\n"
-        markdown_text += f"{' ' * blank_num}|---|---|---|---|---|\n"
+        """
+        gen param md table
+        :param field_dict_list:
+        :param blank_num: table str indent
+        :return:
+            |param name|type|default value|description|other|
+            |---|---|---|---|---|
+            |age|string|10|user age||
+        """
+        blank_num_str: str = " " * blank_num
+        markdown_text: str = f"{blank_num_str}|param name|type|default value|description|other|\n"
+        markdown_text += f"{blank_num_str}|---|---|---|---|---|\n"
         field_dict_list = sorted(field_dict_list, key=lambda x: x["param_name"])
         for field_info_dict in field_dict_list:
             default = field_info_dict["default"]
@@ -64,12 +89,12 @@ class PaitMd(PaitBaseParse):
             type_ = field_info_dict["type"]
             if default is Undefined:
                 type_ = "**`Required`**"
-            description = field_info_dict["description"]
             markdown_text += (
-                f"{' ' * blank_num}|{field_info_dict['param_name']}"
+                f"{blank_num_str}"
+                f"|{field_info_dict['param_name']}"
                 f"|{type_}"
                 f"|{default}"
-                f"|{description}"
+                f"|{field_info_dict['description']}"
                 f"|{field_info_dict['other'] or ''}"
                 f"|\n"
             )
@@ -96,6 +121,7 @@ class PaitMd(PaitBaseParse):
                 elif pait_model.status:
                     status_text = f"{pait_model.status.value}"
 
+                # func or interface details
                 func_code: CodeType = pait_model.func.__code__
                 markdown_text += f"|Author|Status|func|description|\n"
                 markdown_text += f"|---|---|---|---|\n"
@@ -112,7 +138,7 @@ class PaitMd(PaitBaseParse):
                 markdown_text += f"- Method: {','.join(pait_model.method_list)}\n"
                 markdown_text += f"- Request:\n"
 
-                field_dict: Dict[str, List[Dict[str, Any]]] = self._parse_func_param(pait_model.func)
+                field_dict: Dict[str, List[Dict[str, Any]]] = self._parse_func_param_to_field_dict(pait_model.func)
                 # request body info
                 field_key_list: List[str] = sorted(field_dict.keys())
                 for field in field_key_list:
@@ -139,8 +165,7 @@ class PaitMd(PaitBaseParse):
                             markdown_text += f"{' ' * 12}{resp_model.header}\n"
                         if resp_model.response_data:
                             markdown_text += f"{' ' * 8}- Data\n\n"
-                            schema_dict: dict = resp_model.response_data.schema()
-                            field_dict_list = self._parse_schema(schema_dict)
+                            field_dict_list = self._parse_schema(resp_model.response_data.schema())
                             markdown_text += self.gen_md_param_table(field_dict_list, blank_num=12)
                             markdown_text += f"{' ' * 8}- Example Response Data Json\n\n"
                             markdown_text += self.gen_example_json(field_dict_list, blank_num=12)