+++ pait/api_doc/pait_yaml.py
@@ -3,6 +3,7 @@ from typing import Any, Dict, Optional
 import yaml
 
 from pait.model import PaitCoreModel
+
 from .base_parse import PaitBaseParse  # type: ignore